Part of https://github.com/ParkingReformNetwork/reform-map/issues/593. This makes solid improvements to the speed of `PlaceFilterManager.ensureCache`.

However, note that the times were already pretty small to begin with. We need to diagnose what is causing the main slowdown in 593.

## Benchmark

Used this diff:

```diff
diff --git a/src/js/FilterState.ts b/src/js/FilterState.ts
index 90ce96a1..b4223b62 100644
--- a/src/js/FilterState.ts
+++ b/src/js/FilterState.ts
@@ -131,6 +131,7 @@ export class PlaceFilterManager {
     const matchedPolicyRecords: Record<PlaceId, MatchedPolicyRecords> = {};
     const matchedCountries = new Set<string>();
     let numMatchedPolicyRecords = 0;
+    console.time('ensureCache')
     for (const placeId in this.entries) {
       const matchedRecords = this.getMatchingPolicyRecords(placeId);
       if (!matchedRecords) continue;
@@ -141,6 +142,7 @@ export class PlaceFilterManager {
         matchedRecords.reduceMinIdx.length +
         matchedRecords.rmMinIdx.length;
     }
+    console.timeEnd('ensureCache')
 
     this.cache = {
       state: currentState,
```

Before:

[Debug] ensureCache: 1.724ms (localhost, line 44452)
[Debug] ensureCache: 12.185ms (localhost, line 44452)
[Debug] ensureCache: 6.808ms (localhost, line 44452)
[Debug] ensureCache: 12.381ms (localhost, line 44452)
[Debug] ensureCache: 6.007ms (localhost, line 44452)
[Debug] ensureCache: 10.614ms (localhost, line 44452)

After:

[Debug] ensureCache: 1.192ms (localhost, line 44452)
[Debug] ensureCache: 5.217ms (localhost, line 44452)
[Debug] ensureCache: 0.617ms (localhost, line 44452)
[Debug] ensureCache: 10.342ms (localhost, line 44452)
[Debug] ensureCache: 3.109ms (localhost, line 44452)
[Debug] ensureCache: 3.724ms (localhost, line 44452)
[Debug] ensureCache: 3.961ms (localhost, line 44452)
[Debug] ensureCache: 3.750ms (localhost, line 44452)